requirements.txt
pillow
opencv-python
astroalign
numpy
